Make primary and secondary navs easier to distinguish
Add more space between the primary nav (Home, WebSocket, GraphQL, Settings) and the secondary nav (the sections of the current app page, for example Request, Options, and Response).

When I first opened Postwoman and tried to understand the interface, the primary and secondary navs looked like just one big primary nav to me. I clicked on each primary nav item one by one to explore, and I only noticed that secondary nav items worked differently when I clicked on one and it didn’t take me to a new page. There is a tiny dividing line between the navs in the sidebar, but it was so subtle I didn’t see it until after I noticed that the button I clicked had worked differently.

There are many possible ways to make the difference in navs more obvious, such as making the buttons look more different or adding a textual header, but adding spacing seems like a good, simple initial improvement. I chose 16px of extra spacing on each side of the divider because it looked best to my eyes compared to 8px and 12px. 16px is also used in some styles on nearby elements, so it keeps the interface somewhat consistent.