fix: Self-host unable to use Azure oauth
closes #3050

The `passReqToCallback=true` alters the arguments passed to `MicrosoftStrategy.validate` such that first argument is a `request` object and other arguments are shifted across one. This meant that actual values for `accessToken` and critically `profile` weren't in the expected arguments and therefore the rest of the method would break.
I believe `passReqToCallback` was added in error in 81cb0d43d78e41a3ce1e82260a10fa9baedfbc64. This method could never work in it's current form with this option set to true.