refactor: rely on arg mutation in the utility functions
Creating a deep copy and returning the updated value is not possible after the processing happens in the individual scripting API methods. This is because the utility functions are processed at first, and the API methods, when invoked, can't return the updated value to the execution context since the control flow never comes back to the function surroundings.