refactor: issued handles track updates to the resource
By returning `getCollectionHandle` from methods that return a handle, for instance, `createRESTRootCollection`, we could ensure any updates to the respective resource are reflected in the above-issued handle since updates to the store would count as a reactive dependency within `getCollectionHandle` since it accesses the store and thereby signifying a reactive update to `createRESTRootCollection` method which consumes to process the return value. The reactive update wasn't observed while testing, hence relevant TODO comments are added for future investigation.