fix(desktop): csp override for `faraday-cage`
CSP is a web security feature that applies to Tauri apps (via WebKit)
and it controls what resources the page can load and connect to.

When the app tries to connect to the associated WASM module
embedded as a data URI with base64 encoding,
(which starts with `data:application/wasm;base64`),
followed by a string of encoded binary data, the request fails
with `script_fail` since the directive only allows connections
to specific sources defined in `tauri.conf.json`,
also via the `ipc:` URLs, i.e. `http://ipc.localhost`, and
`https://api.hoppscotch.io` for the rest.

Exact error:
```
Refused to connect to data:application/wasm;base64,AGFzbQEAAAABjwd0YA...
Load failed
```

The previous directive didn't include permission for `data:` URLs,
so when the app tries to load the WASM module,
tauri webkit view blocks the connection.

This new directive fixes that by first
including `data:` which allows connections to data URIs,
and second, it adds a wildcard `*` which permits connections
to any other source (good enough for now) for flexibility