fix(desktop): align connected instance with launch
The current implementation uses two separate stores
for the desktop app: one for the `InstanceSwitcherService`
and another for the main desktop app.

This creates a de-sync between the vendored app
launched by the core service, vs the app expected
by the `InstanceSwitcherService`.

This adds a overwrite for `InstanceSwitcherService` store
to sync `"connected"` status correctly.

Closes HFE-894

During the development of the new kernel desktop arch (see HFE-672),
separate stores were introduced as a transitional necessity
to handle initialization order dependencies.

Basically a couple of things including why native interceptor
needed to be ready before `PersistenceService` and
managing storage contexts during web app bundle transitions,
plus eliminating some of the circular service dependencies
that was handled eventually with the `InitializationService`.

This is now being phased out in favor of better platform
and system agnostic path resolution (see HFE-829 and sub ticket HFE-864)
with the goal of eventually introducing a
[kernel](https://github.com/hoppscotch/hoppscotch/tree/main/packages/hoppscotch-common/src/kernel)
store for hoppscotch-desktop,
see HFE-877 for part 6 of the process.